requirements.txt
numpy==1.22
fsspec==2024.10.0
seaborn==0.13.2
deepspeed==0.10.0
Pillow==9.4.0
transformers==4.35.2
diffusers==0.24.0
accelerate==0.21.0
peft==0.13.2
controlnet_aux==0.0.9
Jinja2==3.1.4
huggingface-hub==0.25.2
retrying
setuptools>=40.8.0
open_clip_torch==2.29.0
multilingual_clip==1.0.10
pytorch_lightning==2.1.2